+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Changelog for the docker image are [here](/docker_CHANGELOG.md).
 
 ### Fixed
 
+- avoid context cache collisions by including query shape options in cache keys (#1934)
+- cap DataLoader caches and discarded loader list (#1935)
 - pass context through batch loader paths to prime ContextCache (#1936)
 
 ## [0.2.6]